@@ -1,21 +1,19 @@
-import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
-import Detail from "./routes/Detail";
-import Home from "./routes/Home";
+import { BrowserRouter as Routes, Route } from 'react-router-dom';
+import Main from './componts/Main';
+import Movie from './components/Movie';
+import Detail from './components/MovieDetail';
+import NotFound from './components/NotFound';
+import { MOVIE, ROOT } from './constants/path';
 function App() {
   return (
-    <Router>
-      <Switch>
-        <Route path="/hello">
-          <h1>Hello</h1>
-        </Route>
-        <Route path="/movie">
-          <Detail />
-        </Route>
-        <Route path="/">
-          <Home />
-        </Route>
-      </Switch>
-    </Router>
+    <div className='container'>
+      <Routes>
+        <Route path={ROOT.ROOT} element={<Main />} />
+        <Route path={ROOT.MOVIE} element={<Movie />} />
+        <Route path={MOVIE.VIEW} element={<Detail />} />
+        <Route path={ROOT.NOTFOUND} element={<NotFound />} />
+      </Routes>
+    </div>
   );
 }
 
